multi-language support

+++ speech_dist.sh
@@ -1,6 +1,12 @@
 #!/bin/bash
 
-DISTDIR=data/dist
+if [ $# != 1 ] ; then
+    echo "usage: $0 <lang>"
+    exit 1
+fi
+
+LANG=$1
+DISTDIR=data/dist/$LANG
 
 rm -rf $DISTDIR
 mkdir $DISTDIR
@@ -8,10 +14,10 @@ mkdir $DISTDIR
 datum=`date +%Y%m%d`
 
 #
-# kaldi nnet3 models de
+# kaldi nnet3 models 
 #
 
-AMNAME="kaldi-nnet3-voxforge-de-r$datum"
+AMNAME="kaldi-nnet3-voxforge-${LANG}-r$datum"
 
 mkdir "$DISTDIR/$AMNAME"
 
@@ -22,13 +28,13 @@ function export_kaldi_nnet3 {
 
     mkdir "$DISTDIR/$AMNAME/$EXPNAME"
 
-    cp data/dst/speech/de/kaldi/exp/nnet3/$EXPNAME/final.mdl                  $DISTDIR/$AMNAME/$EXPNAME/
-    cp data/dst/speech/de/kaldi/exp/nnet3/$EXPNAME/cmvn_opts                  $DISTDIR/$AMNAME/$EXPNAME/ 2>/dev/null 
+    cp data/dst/speech/${LANG}/kaldi/exp/nnet3/$EXPNAME/final.mdl                  $DISTDIR/$AMNAME/$EXPNAME/
+    cp data/dst/speech/${LANG}/kaldi/exp/nnet3/$EXPNAME/cmvn_opts                  $DISTDIR/$AMNAME/$EXPNAME/ 2>/dev/null 
 
-    cp data/dst/speech/de/kaldi/exp/nnet3/$GRAPHNAME/HCLG.fst                 $DISTDIR/$AMNAME/$EXPNAME/
-    cp data/dst/speech/de/kaldi/exp/nnet3/$GRAPHNAME/words.txt                $DISTDIR/$AMNAME/$EXPNAME/
-    cp data/dst/speech/de/kaldi/exp/nnet3/$GRAPHNAME/num_pdfs                 $DISTDIR/$AMNAME/$EXPNAME/
-    cp data/dst/speech/de/kaldi/exp/nnet3/$GRAPHNAME/phones/align_lexicon.int $DISTDIR/$AMNAME/$EXPNAME/
+    cp data/dst/speech/${LANG}/kaldi/exp/nnet3/$GRAPHNAME/HCLG.fst                 $DISTDIR/$AMNAME/$EXPNAME/
+    cp data/dst/speech/${LANG}/kaldi/exp/nnet3/$GRAPHNAME/words.txt                $DISTDIR/$AMNAME/$EXPNAME/
+    cp data/dst/speech/${LANG}/kaldi/exp/nnet3/$GRAPHNAME/num_pdfs                 $DISTDIR/$AMNAME/$EXPNAME/
+    cp data/dst/speech/${LANG}/kaldi/exp/nnet3/$GRAPHNAME/phones/align_lexicon.int $DISTDIR/$AMNAME/$EXPNAME/
 
 }
 
@@ -37,14 +43,14 @@ export_kaldi_nnet3 nnet_tdnn_a  nnet_tdnn_a/graph
 
 mkdir "$DISTDIR/$AMNAME/extractor"
 
-cp data/dst/speech/de/kaldi/exp/nnet3/extractor/final.mat            "$DISTDIR/$AMNAME/extractor/"
-cp data/dst/speech/de/kaldi/exp/nnet3/extractor/global_cmvn.stats    "$DISTDIR/$AMNAME/extractor/"
-cp data/dst/speech/de/kaldi/exp/nnet3/extractor/final.dubm           "$DISTDIR/$AMNAME/extractor/"
-cp data/dst/speech/de/kaldi/exp/nnet3/extractor/final.ie             "$DISTDIR/$AMNAME/extractor/"
-cp data/dst/speech/de/kaldi/exp/nnet3/extractor/splice_opts          "$DISTDIR/$AMNAME/extractor/"
-cp data/dst/speech/de/kaldi/exp/nnet3/ivectors_test/conf/splice.conf "$DISTDIR/$AMNAME/extractor/"
+cp data/dst/speech/${LANG}/kaldi/exp/nnet3/extractor/final.mat            "$DISTDIR/$AMNAME/extractor/"
+cp data/dst/speech/${LANG}/kaldi/exp/nnet3/extractor/global_cmvn.stats    "$DISTDIR/$AMNAME/extractor/"
+cp data/dst/speech/${LANG}/kaldi/exp/nnet3/extractor/final.dubm           "$DISTDIR/$AMNAME/extractor/"
+cp data/dst/speech/${LANG}/kaldi/exp/nnet3/extractor/final.ie             "$DISTDIR/$AMNAME/extractor/"
+cp data/dst/speech/${LANG}/kaldi/exp/nnet3/extractor/splice_opts          "$DISTDIR/$AMNAME/extractor/"
+cp data/dst/speech/${LANG}/kaldi/exp/nnet3/ivectors_test/conf/splice.conf "$DISTDIR/$AMNAME/extractor/"
 
-cp data/dst/speech/de/kaldi/RESULTS.txt $DISTDIR/$AMNAME/
+cp data/dst/speech/${LANG}/kaldi/RESULTS.txt $DISTDIR/$AMNAME/
 cp README.md "$DISTDIR/$AMNAME"
 cp LICENSE   "$DISTDIR/$AMNAME"
 cp AUTHORS   "$DISTDIR/$AMNAME"
@@ -62,10 +68,10 @@ popd
 rm -r "$DISTDIR/$AMNAME"
 
 #
-# kaldi gmm models de
+# kaldi gmm models 
 #
 
-AMNAME="kaldi-gmm-voxforge-de-r$datum"
+AMNAME="kaldi-gmm-voxforge-${LANG}-r$datum"
 
 mkdir "$DISTDIR/$AMNAME"
 
@@ -76,15 +82,15 @@ function export_kaldi_model {
 
     mkdir "$DISTDIR/$AMNAME/$EXPNAME"
 
-    cp data/dst/speech/de/kaldi/exp/$EXPNAME/final.mdl   $DISTDIR/$AMNAME/$EXPNAME/
-    cp data/dst/speech/de/kaldi/exp/$EXPNAME/final.mat   $DISTDIR/$AMNAME/$EXPNAME/
-    cp data/dst/speech/de/kaldi/exp/$EXPNAME/splice_opts $DISTDIR/$AMNAME/$EXPNAME/          2>/dev/null
-    cp data/dst/speech/de/kaldi/exp/$EXPNAME/cmvn_opts   $DISTDIR/$AMNAME/$EXPNAME/          2>/dev/null 
-    cp data/dst/speech/de/kaldi/exp/$EXPNAME/delta_opts  $DISTDIR/$AMNAME/$EXPNAME/          2>/dev/null 
+    cp data/dst/speech/${LANG}/kaldi/exp/$EXPNAME/final.mdl   $DISTDIR/$AMNAME/$EXPNAME/
+    cp data/dst/speech/${LANG}/kaldi/exp/$EXPNAME/final.mat   $DISTDIR/$AMNAME/$EXPNAME/
+    cp data/dst/speech/${LANG}/kaldi/exp/$EXPNAME/splice_opts $DISTDIR/$AMNAME/$EXPNAME/          2>/dev/null
+    cp data/dst/speech/${LANG}/kaldi/exp/$EXPNAME/cmvn_opts   $DISTDIR/$AMNAME/$EXPNAME/          2>/dev/null 
+    cp data/dst/speech/${LANG}/kaldi/exp/$EXPNAME/delta_opts  $DISTDIR/$AMNAME/$EXPNAME/          2>/dev/null 
 
-    cp data/dst/speech/de/kaldi/exp/$GRAPHNAME/HCLG.fst  $DISTDIR/$AMNAME/$EXPNAME/
-    cp data/dst/speech/de/kaldi/exp/$GRAPHNAME/words.txt $DISTDIR/$AMNAME/$EXPNAME/
-    cp data/dst/speech/de/kaldi/exp/$GRAPHNAME/num_pdfs  $DISTDIR/$AMNAME/$EXPNAME/
+    cp data/dst/speech/${LANG}/kaldi/exp/$GRAPHNAME/HCLG.fst  $DISTDIR/$AMNAME/$EXPNAME/
+    cp data/dst/speech/${LANG}/kaldi/exp/$GRAPHNAME/words.txt $DISTDIR/$AMNAME/$EXPNAME/
+    cp data/dst/speech/${LANG}/kaldi/exp/$GRAPHNAME/num_pdfs  $DISTDIR/$AMNAME/$EXPNAME/
 
 }
 
@@ -97,7 +103,7 @@ export_kaldi_model tri3b_mpe       tri3b_denlats/dengraph
 export_kaldi_model tri3b_mmi       tri3b_denlats/dengraph
 export_kaldi_model tri3b_mmi_b0.05 tri3b_denlats/dengraph
 
-cp data/dst/speech/de/kaldi/RESULTS.txt $DISTDIR/$AMNAME/
+cp data/dst/speech/${LANG}/kaldi/RESULTS.txt $DISTDIR/$AMNAME/
 cp README.md "$DISTDIR/$AMNAME"
 cp LICENSE   "$DISTDIR/$AMNAME"
 cp AUTHORS   "$DISTDIR/$AMNAME"
@@ -115,17 +121,17 @@ popd
 rm -r "$DISTDIR/$AMNAME"
 
 #
-# cont sphinx model de
+# cont sphinx model
 #
 
-AMNAME="cmusphinx-cont-voxforge-de-r$datum"
+AMNAME="cmusphinx-cont-voxforge-${LANG}-r$datum"
 
 mkdir "$DISTDIR/$AMNAME"
 mkdir "$DISTDIR/$AMNAME/model_parameters"
 
-cp -r data/dst/speech/de/cmusphinx_cont/model_parameters/voxforge.cd_cont_6000 "$DISTDIR/$AMNAME/model_parameters"
-cp -r data/dst/speech/de/cmusphinx_cont/etc "$DISTDIR/$AMNAME"
-cp data/dst/speech/de/cmusphinx_cont/voxforge.html "$DISTDIR/$AMNAME"
+cp -r data/dst/speech/${LANG}/cmusphinx_cont/model_parameters/voxforge.cd_cont_6000 "$DISTDIR/$AMNAME/model_parameters"
+cp -r data/dst/speech/${LANG}/cmusphinx_cont/etc "$DISTDIR/$AMNAME"
+cp data/dst/speech/${LANG}/cmusphinx_cont/voxforge.html "$DISTDIR/$AMNAME"
 cp README.md "$DISTDIR/$AMNAME"
 cp LICENSE   "$DISTDIR/$AMNAME"
 cp AUTHORS   "$DISTDIR/$AMNAME"
@@ -138,17 +144,17 @@ popd
 rm -r "$DISTDIR/$AMNAME"
 
 #
-# ptm sphinx model de
+# ptm sphinx model
 #
 
-AMNAME="cmusphinx-ptm-voxforge-de-r$datum"
+AMNAME="cmusphinx-ptm-voxforge-${LANG}-r$datum"
 
 mkdir "$DISTDIR/$AMNAME"
 mkdir "$DISTDIR/$AMNAME/model_parameters"
 
-cp -r data/dst/speech/de/cmusphinx_ptm/model_parameters/voxforge.cd_ptm_5000 "$DISTDIR/$AMNAME/model_parameters"
-cp -r data/dst/speech/de/cmusphinx_ptm/etc "$DISTDIR/$AMNAME"
-cp data/dst/speech/de/cmusphinx_ptm/voxforge.html "$DISTDIR/$AMNAME"
+cp -r data/dst/speech/${LANG}/cmusphinx_ptm/model_parameters/voxforge.cd_ptm_5000 "$DISTDIR/$AMNAME/model_parameters"
+cp -r data/dst/speech/${LANG}/cmusphinx_ptm/etc "$DISTDIR/$AMNAME"
+cp data/dst/speech/${LANG}/cmusphinx_ptm/voxforge.html "$DISTDIR/$AMNAME"
 cp README.md "$DISTDIR/$AMNAME"
 cp LICENSE   "$DISTDIR/$AMNAME"
 cp AUTHORS   "$DISTDIR/$AMNAME"
@@ -161,27 +167,27 @@ popd
 rm -r "$DISTDIR/$AMNAME"
 
 #
-# srilm de
+# srilm
 #
 
-LMNAME="srilm-voxforge-de-r$datum.arpa"
-cp data/dst/speech/de/srilm/lm.arpa data/dist/$LMNAME
+LMNAME="srilm-voxforge-${LANG}-r$datum.arpa"
+cp data/dst/speech/${LANG}/srilm/lm.arpa data/dist/$LMNAME
 gzip data/dist/$LMNAME
 
 # 
-# cmuclmtk de
+# cmuclmtk
 #
 
-# LMNAME="cmuclmtk-voxforge-de-r$datum.arpa"
-# cp data/dst/speech/de/cmusphinx_cont/voxforge.arpa $DISTDIR/$LMNAME
+# LMNAME="cmuclmtk-voxforge-${LANG}-r$datum.arpa"
+# cp data/dst/speech/${LANG}/cmusphinx_cont/voxforge.arpa $DISTDIR/$LMNAME
 # gzip $DISTDIR/$LMNAME
 
 #
-# sequitur de
+# sequitur
 #
 
-MODELNAME="sequitur-voxforge-de-r$datum"
-cp data/dst/speech/de/sequitur/model-6 $DISTDIR/$MODELNAME
+MODELNAME="sequitur-voxforge-${LANG}-r$datum"
+cp data/dst/speech/${LANG}/sequitur/model-6 $DISTDIR/$MODELNAME
 gzip $DISTDIR/$MODELNAME
 
 #
@@ -196,5 +202,5 @@ cp AUTHORS   "$DISTDIR"
 # upload
 #
 
-echo rsync -avPz --delete --bwlimit=256 data/dist goofy:/var/www/html/voxforge/de
+echo rsync -avPz --delete --bwlimit=256 data/dist/${LANG} goofy:/var/www/html/voxforge/
